+++ app/widget_maker.go
@@ -16,9 +16,10 @@ import (
 	"github.com/wtfutil/wtf/modules/clocks"
 	"github.com/wtfutil/wtf/modules/cmdrunner"
 	"github.com/wtfutil/wtf/modules/covid"
-	"github.com/wtfutil/wtf/modules/cryptoexchanges/bittrex"
-	"github.com/wtfutil/wtf/modules/cryptoexchanges/blockfolio"
-	"github.com/wtfutil/wtf/modules/cryptoexchanges/cryptolive"
+	"github.com/wtfutil/wtf/modules/cryptocurrency/bittrex"
+	"github.com/wtfutil/wtf/modules/cryptocurrency/blockfolio"
+	"github.com/wtfutil/wtf/modules/cryptocurrency/cryptolive"
+	"github.com/wtfutil/wtf/modules/cryptocurrency/mempool"
 	"github.com/wtfutil/wtf/modules/datadog"
 	"github.com/wtfutil/wtf/modules/devto"
 	"github.com/wtfutil/wtf/modules/digitalclock"
@@ -47,7 +48,6 @@ import (
 	"github.com/wtfutil/wtf/modules/kubernetes"
 	"github.com/wtfutil/wtf/modules/logger"
 	"github.com/wtfutil/wtf/modules/mercurial"
-	"github.com/wtfutil/wtf/modules/mempool"
 	"github.com/wtfutil/wtf/modules/nbascore"
 	"github.com/wtfutil/wtf/modules/newrelic"
 	"github.com/wtfutil/wtf/modules/opsgenie"
