+++ modules/cryptocurrency/bittrex/bittrex.go
