+++ modules/cryptocurrency/bittrex/display.go
