+++ modules/cryptocurrency/bittrex/settings.go
