+++ modules/cryptocurrency/bittrex/widget.go
