+++ modules/cryptocurrency/blockfolio/settings.go
