+++ modules/cryptocurrency/blockfolio/widget.go
