+++ modules/cryptocurrency/cryptolive/price/settings.go
