+++ modules/cryptocurrency/cryptolive/settings.go
